Debugged and completed the menu code.
The mode that allows displaying a message before showing available
options has been added, and used in Special Stages for menus that allow
selecting what to do in case of failure.

Pausing in Special Stages is no longer possible when such a menu is
displayed. It would be redundant anyway.

Added confirmation submenus to several options, such as "Restart game".