[Android]Implemented the android-specific control configurations.
* A dedicated button that allows Sonic to transform into his super form
has been added. It only appears if Sonic has all seven Chaos Emeralds,
is not in the final act, either has 50 rings or more or already has his
super form, and only works when the above is true and Sonic is currently
jumping.
* To facilitate the implementation of the super button (see above) the
jump button now sends a signal for a C button, instead of A.
* The on-screen buttons can now be repositioned and their opacity can be
adjusted.
* There are three available sizes for the D-pad and the jump button, to
accomodate tablet users. The other two (pause and super) have fixed sizes.
* The optional gyroscope control scheme for the special stages
can now be configured: the angle of the tilt that is required to simulate
the push of a button, selecting an alternative mode where
tilting the device more and more causes the game to think the player is
button mashing (not analog controls, but... kinda, sorta, somewhat close :) ),
and the configuration that goes with this second mode.
* The virtual controls no longer depend on the default GameMaker Studio
functions, and directly influence the objInput object.
* A mode where the D-Pad appears wherever the user places their thumb
has been implemented.
* The slowdown at the beginning of the game seen when testing on
hardware has been fixed.