Another small round of fixes and enhancements
General:
- The level select menu has been recreated with the new menu system and
rendered compatible with touchscreens.
- The score tally can now be set to three different methods: constants
speed, which is now it is now; variable speed, in which the number of
points counted down every frame is calculated to make the end of the
countdown coincide with the "cash register" chime; and random digits,
which is the same as variable speed, but with actual numbers replaced
with random sets of digits (still need to see whether this is what
VectorSatyr had in mind).
- As part of a small refactoring, both score tallies (end of act and
special stages) now share code pertaining to managing the countdown.
- The "stage complete" music has been reinstated at the end of Special
Stages: now, after the Chaos Emerald or the Time Stone falls from the
sky, the music is cut short by the appropriate sound effect
(respectively, a "chaos emerald" jingle from Sonic 3 & Knuckles or a
"time stone" jingle from Sonic CD) and only then does the score tally
appear, accompanied by the "stage complete" music.
- Fixed a bug where losing to Metal Sonic in Special Stages made both
"Metal Sonic wins" and "Out of Rings" messages appear at once.
- Fixed a graphical oddity when remapping only one key.

Android-specific:
- Touchscreen controls now properly reactivate for the Metal Sonic boss
fight at the start of AAZ act 2.
- The game now can only be played in landscape: upside-down landscape
mode has been disabled, because it made the gyroscope freak out.
- The level select can now be accessed by tapping the screen every time
a new name appears during the startup sequence and then by touching the
screen with two fingers when the game prompts the player to "touch to
start".