30 FPS mode. Most items are good-looking now.
Changed the remaining specific and generic items so that they also look good in 30 FPS mode.
Only the Wind Shield is not OK yet.
Corrected the wonkiness in the screen size menu option and enabled the F4 key again.
Moved the setting that determines the way the score tally counts down to the video options submenu.