Fixed a couple of regressions.
Removed the remnants of the second configuration specific to Android consoles. Ultimately, a simple constant will be used to that effect, since creating a configuration duplicates too many elements (images, settings etc) .
A recent update to the input code broke the touchscreen controls for Android. They have been fixed.
Fixed a bug in key/button mapping where holding for a bit after mapping the A or START buttons would cause the game to try to map these keys again; doing so with B would return to the global options screen.