Started working on integrating surfaces in regular drawing operations.
The end goal is to reduce the resorces used for drawing tiles and parallax backgrounds by using surfaces that pre-combine textures with tiles or backgrounds and drawing them instead, essentially replacing drawing lots of small elements by drawing few big ones.
So far, the system for tiles is working, but the pre-drawing takes some time, which is especially noticeable on Android. Work on the parallax backgrounds started, but has been entirely commented out so far.