Edited the end of Tidal Tubes Act 2 to prevent softlocks.
It could happen that if the player broke the flower capsule at the end while spindashing, he would fall off on the right, and get stuck, softlocking the game. The level has been slightly edited to prevent this, as the platform containing the capsule has been extended to the right.