Changed the act end & special stage score card heading.
Put the heading in the same font as the "ACT" mention on zone name cards.
Special Stage score card is now in a separate screen, to be more in line with
official titles.
TODO: bring back the info on how to become Super Sonic and make sure the fake
card at the end of DDZ reflects these changes.