Added Special Stages to Level Select & Added a No Save Slot
Changed the save system in two ways: it now saves to memory instead of the file "slot 0", and if there is a save file present, everything is shifted to the right. Another slot has been added to compensate.
The Level select menu now allows to select special stages. Unlike the rest, these stages are run in a slightly different mode: upon completing a special stage, the game offers to either try another one or reset the game; there is no way to continue a regular playthrough started with a special stage on level select.

The bug where the level select and the regular menu appeared at once has apparently finally been fixed, but I'll need to run a test on Android to make sure.