First pass: Sprites and Shaders
The purpose of this update to do the simplest thing that will speed up compilation and run times.

Deleted all unused (ie. unreferenced) sprites and cropped the remaining sprites that needed it, adjusting the collision dimensions if necessary. Some of the removed content will probably be added again later, but for now they are not necessary. Also got rid of a couple of unused shaders.