Second pass: Bug fixes (1)
Terrain mapping used a ray sensor that extends from the player's origin to "offset_y + 1" pixels downward in a while loop to detect declining slopes; however, it did not scan the pixels inbetween. Consequently, if the player found terrain that lied between these two points, the terrain mapping sensor would miss it, causing a soft lock. This has been fixed in two ways. First, the mapping sensor is now a box; second, it only loops "offset_y + 1" times every step, instead of indefinitely, so any "mistakes" won't cause a soft-lock.

This was by far the most immediately apparent issue when compiling the game in 1.4.9999, so it was addressed first.