Second pass: improve collision detection on slopes
Update collision code so that slope collision works more smoothly.
Removed leftover collision code that was commented out from previous
commits.

After "Second Pass: Bug Fixes (1)," additional collision symptoms
emerged when interacting with terrain that was slanted or curved. This
is because the collision check in player_collision_floor was not
coordinated with the changes to player_get_terrain and
player_floor_transfer; slope readjustment was changed to use a box,
but the actual acquisition of the terrain used a ray.

Upon further reflection, the original change to player_get_terrain
and player_floor_transfer was incorrect; there was the possibility that
the terrain could be mapped at the player's head area, when it should
only be found at the player's feet. This has also been addressed.

Rather than leave residue of previous attempts at fixing the issues,
any leftover code blocks that were commented out have been expunged
completely. If necessary, they can be restored by copying the relevant
code from the git diffs.