Update player_floor_transfer.gml
Update player_floor_transfer to correct mistakes made in "Second pass:
improve collision detection on slopes".

This reverts some of the changes made to player_floor_transfer in the
previous commit that were causing collision side-effects. It also adds
a proper return statement to the script which was accidentally left out.