Second pass: fix Galacnik projectile behavior
Galacnik's fireball would not properly track one of the falling
platforms during the Planetary Panic boss fight, resulting in a soft
lock. This was due to the relationship between the Galacnik object that
fires the projectile (objGalacnikSmall) and the projectile itself
(objGalacnikFireball.) There was temporal coupling that got exposed
when the asset tree was rearranged in "First pass: Objects and Rooms",
as doing so changed the order in which each object's step event was
executed.

"Fixed Galacnik fireball behavior #58" attempted to address this issue
by forcing the correct order of execution to occur. The way in which
this was done meant that the fireball could not properly exist unless
the Galacnik object was still present, so code was added to destroy the
former when the latter left the room. This resulted in incorrect
behavior, as originally the fireball could continue to exist and track
its target even as the Galacnik object left the room.

This commit refactors the code for both objects, as well as
objBossPlatform, not only to reduce the temporal coupling but also to
increase the objects' modularity, without affecting the boss'
intended behavior.