Integrated submission for time travel shader
Thanks to @nkrapivin, the time travel shader is now cross-platform! Tested out at least with Android (on my end) and the original code, before my integration and slight trefactoring, has been confirmed to work with the PS4 (on his end) . Made sure the Windows build didn't get corrupted neither.

With this, I removed the "NoShader" objects, as well as the configuration-dependent constant "is TT shader supported". I also used the opportunity to refactor the shader init and draw code that was copy-pasted across three different objects, and make two scripts out of them.