Integrated the new Android library
Also, corrected a crash in the menu triggered by alternatively using the gamepad and the touchscreen in the menu, which made it possible to activate items that weren't exactly items (title or a cursor value out of bounds).

Updated the default fonts to full ASCII sheets, in part to make sure the gamepad's name appears correctly.

Left to do: save and load options, right now they revert to default on every launch.