Added the hardware mapping generator
Added the source for an app that allows to easily generate usable hardware mapping CSV files.

Replaced the manually created ones for Joy-Cons with those generated with the app, as well as added one for XBox controllers.

Also, fixed a bug where mapping two devices in-game while one is already connected would fail to refresh the displayed name in the menu.