Added shield icons #96
After a couple of discussions with @VectorSatyr , here's how it works:
 - In regular acts, the fourth line like in the mockup appears. This was done to keep the information on the currently equipped shield visible if the player gets an invincibility or goes Super, also to help new players realize the rock shield is in fact a shield.
 - In Special Stages, a Super Mario World-like box appears above, turning half-transparent if no shield is obtained. Hopefully the fact that it appears differently in Special Stages will be enough of a top-off that shields in Special Stages 1) are rewards that can be kept after the stage is completed or forfeited and 2) don't work as shields during Special Stages themselves.

TODO: see whether the default position of the Pause button on Android overlaps with the box in Special Stages. If it does, move the box to the side, but again, only on Android, and only if the touchscreen is used.