Started refactoring some duplicated code in HUDs
The handling of the differences between the regular HUD and the one for mobile versions are produced by duplicating some code. This is an attempt to reduce the amount fo duplicated code.

Also, moved the shield icon to the side in Special Stages on Android.

TODO: the progress bar in special stages is still duplicated