Added a handling of incomplete key mappings #85
Exiting the mapping menu while at least one key is unassigned displays a prompt that forces to either remap everything or reset to default.

If the game is reset at this point, the game will check whether all keys are assigned and revert  to default if it isn't the case.