Fixed handling of save file 0 #107
The game no longer renames all save files if a save 0 is detected. No Save is now slot -1. Level select is now slot -1 if a regular Act is selected and -2 if a Special Stage is selected.

The cursor on the save screen was used instead of the save slot number pretty much everywhere. This shouldn't be the case anymore.