Corrected Squibot laser collision #29
Enabling precise collision checks to the sprite solved the issue of it hitting Sonic despite clearly missing. The Egg Walker laser (that uses the same sprite) has been tested and doesn't seem to have been negatively impacted by the change.

The calculation of the angle has also been modified, since with the current formula, the laser would shoot slightly below the player character, making it possible to ignore the laser entirely simply by standing still in front of the badnik. The new coordinates are the ones used for the laser's origin once it's fired.