Final fixes & improvements to touchscreen controls: 1st pass #83
As of right now, touchscreen controls appear when they are not needed, especially when score cards appear.

This first commit is the beginning of the correct implementation of on-screen controls: touchscreen controls appear only when they are needed/active and disappear otherwise.

A rather welcome (and intended) side effect of this is also disabling menus during score cards even on the PC version, which would among other things allow messing with localization options, introducing an edge case of translation not being applied to the big text after it is shown, an edge case of very little interest.