Added another version of title cards #76
Depending on the chosen locale, in some languages (such as Russian) the word "zone" in a title card appears first instead of last, and as such should be written in a different font.

A flag "titlecard_zone_is_first" in translation JSON files is used to determine whether this is the case.