Special Stage optimizations
Disabed drawing shadows on instances that are further than 1000 pixels from the player.

Made sure instances that are out of frame are not visible/drawn except for Metal Sonic to reduce the number of calls to very performance-heavy D3D draw functions.

Thanks to this, the profiler in Special Stage Past 7 shows that the time taken to draw respectively rings, spikeballs and bumpers goes from respectively 20%, 17% and 14,5% of the total execution time to 11%, 5,5% and 8%.