Touched up the plane's behavior to account for Super Sonic #129
Its X is now clamped instead of relying on character speed (which produces the same result, but accounting for what happens when the player gets knocked back repeatedly while running into an enemy.

As the player can still phase through the plane while hitting Big Bomb Badniks, the plane how has an emergency code, repositioning itself when this happens. This is not the most elegant solution, but at least it prevents a softlock.